requirements.txt
requests>=2.31.0
pystray>=0.19.5
Pillow>=10.4.0
schedule>=1.2.1
customtkinter>=5.2.2
